+++ src/control_plane/migration/mysql/mod.rs
@@ -1,0 +1,5 @@
+mod backup_mysql_database;
+mod mysql_backup_options;
+
+pub(crate) use backup_mysql_database::backup_mysql_database;
+pub(crate) use mysql_backup_options::MySqlBackupOptions;